Fix fn files; Update README; Add .png
Moved getLattice filter on stepHex and stepLife out of the conv2 call as it is not
required in the neighbourhood computation; only at the beginning of the
loop are they required, but kept them in the file anyway for consistency
of the main solutions.m file. Moved h and w extraction to top of file as
the getLattice needs it. Moved this h,w extraction procedure to top for
all function files for consistency

Add comments to plotGrid

README sections completely covered, though not yet proofread much.

Tweaked comments and command window message on fig2png.m

Add FinalStates.png to demonstrate procedure at end of solutions.m
script